subindo api para pedido de compra

+++ src/main/java/br/com/api/sales/java/repository/ProductJpaRepository.java
@@ -1,7 +1,5 @@
 package br.com.api.sales.java.repository;
 
-import java.math.BigDecimal;
-
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -9,5 +7,5 @@ import br.com.api.sales.java.model.Product;
 
 public @Repository interface ProductJpaRepository extends JpaRepository<Product, Long> {
 
-	Product findByNameAllIgnoreCaseAndPrice(String name, BigDecimal price);
+	Product findByNameAllIgnoreCaseAndPrice(String name, Double price);
 }
